+++ src/Sentry.Unity.Editor/ConfigurationWindow/EnrichmentTab.cs
@@ -121,6 +121,47 @@ namespace Sentry.Unity.Editor.ConfigurationWindow
                 options.ScreenshotCompression, 1, 100);
 
             EditorGUILayout.EndToggleGroup();
+
+            EditorGUILayout.Space();
+            EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
+            EditorGUILayout.Space();
+            {
+                options.AttachViewHierarchy = EditorGUILayout.BeginToggleGroup(
+                    new GUIContent("Attach Hierarchy", "Try to attach the current scene's hierarchy."),
+                    options.AttachViewHierarchy);
+
+                options.MaxViewHierarchyRootObjects = EditorGUILayout.IntField(
+                    new GUIContent("Max Root GameObjects", "Maximum number of captured GameObjects in " +
+                                                    "a scene root." +
+                                                    "\nDefault: 100"),
+                    options.MaxViewHierarchyRootObjects);
+                if (options.MaxViewHierarchyRootObjects <= 0)
+                {
+                    options.MaxViewHierarchyRootObjects = 0;
+                }
+
+                options.MaxViewHierarchyObjectChildCount = EditorGUILayout.IntField(
+                    new GUIContent("Max Child Count Per Object", "Maximum number of child objects " +
+                                                    "captured for each GameObject." +
+                                                    "\nDefault: 20"),
+                    options.MaxViewHierarchyObjectChildCount);
+                if (options.MaxViewHierarchyObjectChildCount <= 0)
+                {
+                    options.MaxViewHierarchyObjectChildCount = 0;
+                }
+
+                options.MaxViewHierarchyDepth = EditorGUILayout.IntField(
+                    new GUIContent("Max Depth", "Maximum depth of the hierarchy to capture. " +
+                                                    "For example, setting 1 will only capture root GameObjects." +
+                                                    "\nDefault: 10"),
+                    options.MaxViewHierarchyDepth);
+                if (options.MaxViewHierarchyDepth <= 0)
+                {
+                    options.MaxViewHierarchyDepth = 0;
+                }
+
+                EditorGUILayout.EndToggleGroup();
+            }
         }
     }
 }