+++ src/Sentry.Unity.Editor/SentryCli.cs
@@ -35,26 +35,37 @@ namespace Sentry.Unity.Editor
             return propertiesFile;
         }
 
-        public static string SetupSentryCli()
+        public static string SetupSentryCli(string? outDirectory = null, RuntimePlatform? actualBuildHost = null)
         {
-            var sentryCliPlatformName = GetSentryCliPlatformName();
-            var sentryCliPath = GetSentryCliPath(sentryCliPlatformName);
-            SetExecutePermission(sentryCliPath);
+            var executableName = GetSentryCliPlatformExecutable(actualBuildHost);
+            var sentryCliPath = GetSentryCliPath(executableName);
 
+            if (outDirectory is not null)
+            {
+                if (!Directory.Exists(outDirectory))
+                {
+                    throw new DirectoryNotFoundException($"Output project directory not found: {outDirectory}");
+                }
+
+                var outCliPath = Path.Combine(outDirectory, executableName);
+                File.Copy(sentryCliPath, outCliPath, true);
+                sentryCliPath = outCliPath;
+            }
+
+            SetExecutePermission(sentryCliPath);
             return sentryCliPath;
         }
 
-        internal static string GetSentryCliPlatformName(IApplication? application = null)
+        internal static string GetSentryCliPlatformExecutable(RuntimePlatform? buildHost = null)
         {
-            application ??= ApplicationAdapter.Instance;
+            buildHost ??= ApplicationAdapter.Instance.Platform;
 
-            return application.Platform switch
+            return buildHost switch
             {
                 RuntimePlatform.WindowsEditor => SentryCliWindows,
                 RuntimePlatform.OSXEditor => SentryCliMacOS,
                 RuntimePlatform.LinuxEditor => SentryCliLinux,
-                _ => throw new InvalidOperationException(
-                    $"Cannot get sentry-cli for the current platform: {Application.platform}")
+                _ => throw new InvalidOperationException($"Cannot get sentry-cli for {buildHost}")
             };
         }
 
@@ -84,26 +95,6 @@ namespace Sentry.Unity.Editor
             {
                 throw new UnauthorizedAccessException($"Failed to set permission to {filePath}");
             }
-        }
-
-        internal static void AddExecutableToXcodeProject(string projectPath, IDiagnosticLogger? logger)
-        {
-            var executableSource = GetSentryCliPath(SentryCliMacOS);
-            var executableDestination = Path.Combine(projectPath, SentryCliMacOS);
-
-            if (!Directory.Exists(projectPath))
-            {
-                throw new DirectoryNotFoundException($"Xcode project directory not found at {executableDestination}");
-            }
-
-            if (File.Exists(executableDestination))
-            {
-                logger?.LogDebug("sentry-cli executable already found at {0}", executableDestination);
-                return;
-            }
-
-            File.Copy(executableSource, executableDestination);
-            SetExecutePermission(executableDestination);
         }
 
         internal static string? UrlOverride(string? dsnOption, string? urlOverrideOption)