+++ test/Scripts.Integration.Test/integration-test.ps1
@@ -4,8 +4,8 @@
 # └───────────────────────────────────────────────────┘ #
 
 param(
-    [string] $UnityVersion,
-    [string] $Platform,
+    [Parameter(Mandatory = $true)][string] $UnityVersion,
+    [Parameter(Mandatory = $true)][string] $Platform,
     [switch] $Clean,
     [switch] $Repack,
     [switch] $Recreate,
@@ -60,7 +60,13 @@ If (-not(Test-Path -Path $PackageReleaseOutput) -Or $Repack)
     ./test/Scripts.Integration.Test/extract-package.ps1
 }
 
-If (-not(Test-Path -Path "$NewProjectPath") -Or $Recreate)
+# Support recreating the integration test project without cleaning the SDK build (and repackaging).
+if ($Recreate -and (Test-Path -Path $NewProjectPath))
+{
+    Remove-Item -Path $NewProjectPath -Recurse -Force -Confirm:$false
+}
+
+If (-not(Test-Path -Path "$NewProjectPath"))
 {
     Write-Host "Creating Project"
     ./test/Scripts.Integration.Test/create-project.ps1 "$UnityPath"
@@ -76,12 +82,12 @@ If ($Rebuild -or -not(Test-Path -Path $buildDir))
 {
     Write-Host "Building Project"
 
-    If ($Platform -eq "iOS")
+    If (("iOS", "Android-Export") -contains $Platform)
     {
         ./test/Scripts.Integration.Test/build-project.ps1 -UnityPath "$UnityPath" -UnityVersion $UnityVersion -Platform $Platform
-        ./scripts/smoke-test-ios.ps1 Build -IsIntegrationTest -UnityVersion $UnityVersion
+        & "./scripts/smoke-test-$($Platform -eq 'iOS' ? 'ios' : 'android').ps1" Build -IsIntegrationTest -UnityVersion $UnityVersion
     }
-    Else 
+    Else
     {
         ./test/Scripts.Integration.Test/build-project.ps1 -UnityPath "$UnityPath" -CheckSymbols -UnityVersion $UnityVersion -Platform $Platform
     }
@@ -95,9 +101,9 @@ Switch -Regex ($Platform)
     {
         ./test/Scripts.Integration.Test/run-smoke-test.ps1 -Smoke -Crash
     }
-    "^Android$"
+    "^(Android|Android-Export)$"
     {
-        ./scripts/smoke-test-droid.ps1 -IsIntegrationTest
+        ./scripts/smoke-test-android.ps1 -IsIntegrationTest
     }
     "^iOS$"
     {